+++ doc/release-notes.md
@@ -11,12 +11,14 @@ SPDX-License-Identifier: CC-BY-4.0
 ### New Features and Major Changes
 
 * [Design validation for electricity sector in sector networks PR #70](https://github.com/pypsa-meets-earth/pypsa-earth-status/pull/70)
-*
+
+* [Add multi-scenario health status validation reports with ember dataset source PR #61](https://github.com/pypsa-meets-earth/pypsa-earth-status/pull/61)
+
 * [Add electricity generation validation based on IRENA data PR #63](https://github.com/pypsa-meets-earth/pypsa-earth-status/pull/63)
 
-* [Add optional custom subfolder to resources and results folders PR #62](https://github.com/SPSUnipi/pypsa2smspp/pull/62)
+* [Add optional custom subfolder to resources and results folders PR #62](https://github.com/pypsa-meets-earth/pypsa-earth-status/pull/62)
 
-* [Link year and country parameters to build_reference_statistics rule PR #58](https://github.com/SPSUnipi/pypsa2smspp/pull/58)
+* [Link year and country parameters to build_reference_statistics rule PR #58](https://github.com/pypsa-meets-earth/pypsa-earth-status/pull/58)
 
 * [Decompose clean_data into carrier-source-specific rules PR #47](https://github.com/SPSUnipi/pypsa2smspp/pull/47)
 
@@ -29,6 +31,8 @@ SPDX-License-Identifier: CC-BY-4.0
 * [Add functionality to plot OSM electrical grid network plus circuit values and csv export of osm grid line voltages and their lengths PR #38](https://github.com/pypsa-meets-earth/pypsa-earth-status/pull/38)
 
 ### Minor Changes and bug-fixing
+
+* [Adjust a high-level project description PR #71](https://github.com/pypsa-meets-earth/pypsa-earth-status/pull/71)
 
 * [Make validation outputs configuration-specific to avoid stale results when switching validation settings PR #62](https://github.com/pypsa-meets-earth/pypsa-earth-status/pull/62)